fix: enable chat history and session resume in activity chat
Extract useSessionResume hook from duplicated logic in usePRChat and
useActivityChat. Add missing setPRContext and registerResumeSession
calls so history list renders and session selection works from
review/comment chats.
